tweak(tools): apply suggestions from code review
